cfo_agent/BACKUP_20251018_201421/requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# LangChain & LangGraph (compatible versions)
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Environment & Config
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Utilities
python-multipart>=0.0.6

# Streamlit UI
streamlit>=1.30.0
requests>=2.31.0
